# go/model_project_info.go
/*
 * sayhi web service
 *
 * A polite web service that always says hello
 *
 * API version: 1.0.0
 * Contact: marco.micera+sayhi@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package sayhi

import (
	"time"
)

// Project information, including the git hash and the name of the project
type ProjectInfo struct {
	GitHash string `json:"gitHash"`

	ProjectName string `json:"projectName"`

	QueryTime *time.Time `json:"queryTime,omitempty"`
}